+++ src/lib/Components/Consignments/Components/__tests__/ArtworkConsignmentTodo-tests.tsx
@@ -1,22 +1,42 @@
+import { ConsignmentSetup } from "lib/Components/Consignments"
+import ConsignmentTODO from "lib/Components/Consignments/Components/ArtworkConsignmentTodo"
 import * as TODOStories from "../../__stories__/Todo.story"
 import storyRunner from "./Runner"
 
 storyRunner("TODO states:", TODOStories)
 
-describe("User Interaction", () => {
-  // This test is going to end up brittle
-  // see: https://github.com/artsy/emission/issues/577
-
-  it.skip("calls goto Artist on to the artist button", () => {
-    // const testFunc = jest.fn()
-    // const renderer = ReactTestUtils.createRenderer()
-    // renderer.render(<TODO goToArtist={testFunc} />)
-    // const e = renderer.getRenderOutput()
-    // const d = ReactTestUtils.renderIntoDocument(<TODO goToArtist={testFunc} />)
-    // const a = ReactTestUtils.findAllInRenderedTree(e, c => true)
-    // console.log(a)
-    // e.props.children[1].props.onPress()
-    // expect(testFunc).toBeCalled()
-    // console.log(e.props.children[1])
-  })
+it("requires the same metadata props as force", () => {
+  const requiredProps: ConsignmentSetup = {
+    metadata: {
+      category: "a",
+      categoryName: "A",
+      title: "Work",
+      year: "123",
+      medium: "123",
+      width: "123",
+      height: "12",
+      depth: 12,
+      unit: "CM",
+      displayString: "A work",
+    },
+  }
+  const todo = new ConsignmentTODO(requiredProps)
+  expect(todo.canSubmitMetadata(requiredProps)).toBeTruthy()
+})
+it("fails if the right metadata props aren't present", () => {
+  const requiredProps: any = {
+    metadata: {
+      category: "a",
+      categoryName: "A",
+      title: "Work",
+      year: "123",
+      width: "123",
+      height: "12",
+      depth: 12,
+      unit: "CM",
+      displayString: "A work",
+    },
+  }
+  const todo = new ConsignmentTODO(requiredProps)
+  expect(todo.canSubmitMetadata(requiredProps)).toBeFalsy()
 })